move StorageHub to storagehub, Mailboat to root
