+++ mailboat/__init__.py
@@ -1,68 +1,161 @@
-from dataclasses import dataclass
-from typing import Optional
-from mailboat.usrsys.usr import MAILBOX_DEFAULT_SETTING, MailBoxRecord, UserRecord
-from .usrsys.mailbox import MailBox
-from .mailstore import MailStore
-from .usrsys.storage import (
-    MailBoxRecordStorage,
-    MailRecordStorage,
-    ProfileRecordStorage,
-    UserRecordStorage,
-    TokenRecordStorage,
-)
+# Copyright (C) 2021 The Mailboat Contributors
+#
+# This file is part of Mailboat.
+#
+# Mailboat is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Mailboat is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Mailboat.  If not, see <http://www.gnu.org/licenses/>.
+
+from email.message import EmailMessage
+from .usrsys.usr import UserRecord
+from .usrsys.auth import AuthProvider, AuthRequest
+from typing import Any, Dict, List, Optional
+
+from aiosmtpd.smtp import AuthResult, LoginPassword, SMTP
+from .mta import TransferAgent
+from .storagehub import StorageHub
 from unqlite import UnQLite
-from .utils.storage import CommonStorage, UnQLiteStorage
 
 
-class StorageHub(object):
-    """All storages about mailboat"""
+class Mailboat(object):
+    """The entry of Mailboat. This class stores configuration and tools to keep other components running.
 
-    def __init__(self, database: UnQLite) -> None:
-        self.database = database
+    Mailboat splits its feature units as reusable components. Current in-design components include:
+
+    - User System (`mailboat.usrsys`)
+    - Mail Transfer Agent (`mailboat.mta`)
+    - Mail User Agent (in developing)
+    - HTTP API Gateway (in developing)
+
+    This class is also provided as a bridge between different components.
+
+    .. caution:: Though many properties could be changed in runtime, be notice on the side effect!
+    """
+
+    def __init__(
+        self,
+        *,
+        hostname: str,
+        mydomains: List[str],
+        database_path: str,
+        smtpd_port: Optional[int] = None,
+        auth_require_tls: bool = True,
+    ) -> None:
+        if not smtpd_port:
+            smtpd_port = 8025
+        self.mydomains = mydomains
+        """`List[str]`. The domains which should be managed by this instance.
+        For example, the doamin in email address: "random.one@foo.bar" will be seen as local mail address if "foo.bar" in mydomains,
+        all messages sent to the address will be processed as local delivering."""
+        self.hostname = hostname
+        """`str`. The hostname of this server. Don't confuse it to `mydomains`, it's the name of this server.
+        Typically it's the default domain for this instance."""
+        self.database_path: str = database_path
+        """`str`. The path to database"""
+        self.database = UnQLite(database_path)
+        """Database instance. Notice that this property may not be avaliable in future."""
+        self.storage_hub = StorageHub(self.database)
+        """`mailboat.StorageHub`. The references to all storages in mailboat."""
+        self.transfer_agent = TransferAgent(
+            mydomains=mydomains,
+            local_delivery_handler=self.handle_local_delivering,
+            database=self.storage_hub.database,
+            smtpd_auth_handler=self.handle_smtpd_auth,
+            hostname=self.hostname,
+            self_name="transfer_agent.{}".format(self.hostname),
+            smtpd_port=smtpd_port,
+            auth_require_tls=auth_require_tls,
+        )
+        """`mailboat.mta.TransferAgent`. The transfer agent for this instance."""
+        self.auth_provider = AuthProvider(
+            self.storage_hub.user_records, self.storage_hub.token_records
+        )
+        """`mailboat.usrsys.auth.AuthProvider`. The auth provider for this instance."""
         super().__init__()
 
-    def get_common_storage(self, name: str) -> CommonStorage:
-        return UnQLiteStorage(self.database, name)
+    @property
+    def smtpd_port(self):
+        "Mail Transfer Agent's smtpd port."
+        return self.transfer_agent.smtpd_port
 
     @property
-    def user_records(self) -> UserRecordStorage:
-        return UserRecordStorage(self.get_common_storage("users"))
+    def auth_require_tls(self) -> bool:
+        """This property defines should the SMTP Server enable the auth module for non-TLS connections."""
+        return self.transfer_agent.auth_require_tls
 
-    @property
-    def profile_records(self) -> ProfileRecordStorage:
-        return ProfileRecordStorage(self.get_common_storage("profiles"))
+    async def handle_smtpd_auth(
+        self, server: SMTP, method: str, data: Any
+    ) -> AuthResult:
+        """This function is to handle authentication requests from SMTP server.
 
-    @property
-    def mailbox_records(self) -> MailBoxRecordStorage:
-        return MailBoxRecordStorage(self.get_common_storage("mailboxs"))
+        Current supported methods:
+        - login
+        - plain
 
-    @property
-    def mail_records(self) -> MailRecordStorage:
-        return MailRecordStorage(self.get_common_storage("mail_records"))
+        Related:
 
-    @property
-    def mailstore(self) -> MailStore:
-        return MailStore(self.get_common_storage("mails"))
+        - `mailboat.mta.TransferAgent`
+        """
+        if method == "login" or method == "plain":
+            assert isinstance(data, LoginPassword)
+            username: bytes = data.login
+            password: bytes = data.password
+            auth_request = AuthRequest(
+                username=username.decode("utf-8"), password=password.decode("utf-8")
+            )  # TODO (rubicon): support the other charsets
+            result = await self.auth_provider.auth(auth_request)
+            return AuthResult(success=result.success, handled=result.handled)
+        else:
+            return AuthResult(success=False, handled=False)
 
-    @property
-    def token_records(self) -> TokenRecordStorage:
-        return TokenRecordStorage(self.get_common_storage("tokens"))
+    async def handle_local_delivering(self, message: EmailMessage):
+        """This function is to handle local delivering (emails sent to local mail addresses), mostly from the mail transfer agent.
 
-    async def get_mailbox(self, boxid: str) -> Optional[MailBox]:
-        if not boxid:
-            return None
-        record = await self.mailbox_records.find_one({"identity": boxid})
-        if not record:
-            return None
-        return MailBox(record, self.mail_records, self.mailstore, self.mailbox_records)
+        Related:
 
-    async def create_user(self, username: str, password: bytes) -> UserRecord:
-        profile = await self.profile_records.create_new_profile()
-        user = await self.user_records.create_new_user(
-            username, password, profile.identity
+        - `mailboat.mta.TransferAgent`
+        """
+        delivered_to = message["delivered-to"]
+        raise NotImplementedError
+        # TODO (rubicon): complete local delivering
+
+    def start(self):
+        """Start the engine!
+
+        Related:
+
+        - `mailboat.mta.TransferAgent.start`
+        """
+        self.transfer_agent.start()
+
+    def stop(self):
+        """Stop the mailboat instance.
+
+        Related:
+
+        - `mailboat.mta.TransferAgent.destory`
+        """
+        self.transfer_agent.destory()
+
+    async def new_user(
+        self, username: str, nickname: str, email_address: str, password: str
+    ) -> UserRecord:
+        """Create a new user. This method is used for programmaic uses from outside, it does access the storage directly.
+        For internal uses please turn to `mailboat.StorageHub.create_user`.
+        """
+        user = await self.storage_hub.create_user(username, password.encode("ascii"))
+        user.nickname = nickname
+        user.email_address = email_address
+        await self.storage_hub.user_records.update_one(
+            {"profileid": user.profileid}, user
         )
-        for name in MAILBOX_DEFAULT_SETTING:
-            mailbox = await self.mailbox_records.create_mailbox()
-            user.mailboxes[name] = mailbox.identity
-        await self.user_records.update_one({"profileid": user.profileid}, user)
         return user
